+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 -  **BREAKING** Handlers must return `Plan`, not `Message`
+-  **BREAKING** Messages can only contain a string value, not JSON
+
 
 [Unreleased]: https://github.com/atomist/rug/compare/0.20.0...HEAD